+++ src/domain/card/assets/Mastercard.png
