+++ src/domain/card/assets/Visa.png
