+++ src/domain/card/utils/cardInfo.ts
